Move ReconfigDetectionSQLQueryCommand back
This seemed to be unnecessary during rebalance code-path.
